Update to ver 1.5
Added support for AD integrations, script now pulls from the AD server to find the locations (located in the computer object's description) to populate the appropriate field. The script connects to the AD server that the computer is currently using.

Added support for Offline checking. If a register is known to be offline change the description to begin with "offline" and the script will not fill out a form for it during the form filler. User is informed of all registers not being included by a text prompt before form filler is launched. 

Implemented multi threading to speed up the opening of new windows in the form filler. Improves speed by approximately 3x. Users may notice that each page takes longer to fill in, but they all fill in concurrently, reducing time spent waiting on loading.

Various housekeeping and bugfix changes.