Update to 1.8
Added functionality to detect Low hard drive space and replication status failures on the Culinary registers and make the appropriate tickets. Fixed several grammatical errors in the script and changed the phrasing on several items in order to accommodate the new changes. Added additional status types: "HDD problem" and Repl problem" for the corresponding errors. Improved stability of the form filler application by limiting the script to having 10 open chrome windows at one time. for any runs that feature less than 10 tickets, the program will run as normal. If more than 10 tickets are needed for a given run, the user will be warned that opening many chrome windows will lead to instability and directed on how to proceed. The script will require the user to press enter in order to indicate they are ready for the next set of 10 chrome windows after each set until all tickets are exhausted.