Update 2.0
Redesigned the main parsing portion of the code to improve readability and performance.
Fixed a bug where the last item in an email would not be processed if it was offline.
Redefined the error state for repl errors to only show on status 6 as opposed to not 1 or 2
redefined regex commands to be more general but still target only the appropriate registers. This should prevent changes in naming conventions in the future from causing problems and fix a few edge cases with current MPR registers.